Move namespace quota param to model package
